revamped multi-language support
With this change, language files are ordinary Lisp files that can be
loaded to provide language-specific messages, allowing us to avoid
messing about with pathnames.

It also contains some minor features like falling back to English (or
ultimately, Symbolish) if no message is defined for the current
language.

Documentation will be updated in a future commit.